+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.1.0 (2024-12-05)
+
+- Support `config-file` option
+- Support for multi-argument input files
+- Add missing options
+
 ## 1.0.0 (2024-12-03)
 
 - Initial release